# api/openapi.yaml
openapi: 3.0.3
info:
  title: ORACUL API
  version: 0.2.1
  description: |
    Contract between the Spring Boot backend and the Angular frontend. Both sides are generated from this file.
    Every request is tied to an anonymous browser session (HttpOnly cookie ORACUL_SID, created by the backend on the
    first /api request). Data of another session is never visible: it answers 404 like a missing resource.
    Every non-2xx response body is an ApiError { code, message }. No response ever contains a ChatGPT credential.
...
paths:
...
  /api/runs:
    post:
      tags: [runs]
      operationId: startRun
      summary: GENERATE THE FUTURE — snapshot the configuration into a new Generation Run and start the pipeline
      description: |
        Check order: body validation → ChatGPT connection → single active run. The first body violation wins, in the
        order JSON syntax → realism → darkness → optimism → horizon → wildcards → customWildcards → output
        (messages in scenario-panel.md). realism/darkness/optimism must be JSON integers (no strings or fractions);
        horizon codes are case-sensitive. Unknown properties are ignored. Nothing is stored when a check fails.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScenarioConfiguration'
      responses:
        '202':
          description: Run accepted and started asynchronously (status QUEUED or RUNNING)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenerationRun'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/ChatGptRequired'
        '403':
          $ref: '#/components/responses/PlanNotEligible'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'
    get:
      tags: [history]
      operationId: listRecentRuns
      summary: Recent futures — the 20 newest COMPLETED runs of this session, newest first
      responses:
        '200':
          description: Recent futures
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecentRunList'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /api/runs/{runId}/alternatives:
    post:
      tags: [runs]
      operationId: startAlternativeRun
      summary: ALTERNATIVE FUTURE — new run with the same configuration and the same Evidence Pack, different causal path
      parameters:
        - $ref: '#/components/parameters/RunId'
      responses:
        '202':
          description: Alternative run accepted (kind ALTERNATIVE, same evidencePackId as the parent)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenerationRun'
        '401':
          $ref: '#/components/responses/ChatGptRequired'
        '403':
          $ref: '#/components/responses/PlanNotEligible'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  parameters:
    RunId:
      name: runId
      in: path
      required: true
      description: Run id (UUID). A malformed id or a run of another session answers 404 RUN_NOT_FOUND.
      schema:
        type: string
        format: uuid
  responses:
    BadRequest:
      description: Validation failed (code VALIDATION_FAILED); message names the first violated rule
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          examples:
            darkness:
              value: { code: VALIDATION_FAILED, message: darkness must be between 1 and 10 }
            optimism:
              value: { code: VALIDATION_FAILED, message: optimism must be between 1 and 10 }
            realism:
              value: { code: VALIDATION_FAILED, message: realism must be between 1 and 10 }
            horizon:
              value: { code: VALIDATION_FAILED, message: unknown horizon }
            invalidJson:
              value: { code: VALIDATION_FAILED, message: Request body is not valid JSON }
    ChatGptRequired:
      description: CHATGPT_NOT_CONNECTED or CHATGPT_SESSION_EXPIRED
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          examples:
            notConnected:
              value: { code: CHATGPT_NOT_CONNECTED, message: Connect ChatGPT to generate }
            expired:
              value: { code: CHATGPT_SESSION_EXPIRED, message: ChatGPT session expired — please reconnect }
    PlanNotEligible:
      description: CHATGPT_PLAN_NOT_ELIGIBLE
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    NotFound:
      description: RUN_NOT_FOUND
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Conflict:
      description: RUN_ALREADY_ACTIVE, RUN_NOT_COMPLETED, RESEARCH_NOT_READY, EVIDENCE_PACK_NOT_READY, SCENARIO_NOT_READY or RESULT_NOT_READY
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    InternalError:
      description: Last-resort INTERNAL_ERROR without internals (no stack trace, no provider body, no token)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
  schemas:
...
    ApiError:
      type: object
      additionalProperties: false
      required: [code, message]
      properties:
        code:
          type: string
          description: Stable machine code, e.g. VALIDATION_FAILED, RUN_ALREADY_ACTIVE
        message:
          type: string
          description: User-facing English message; never contains internals or credentials

    # ---------- scenario ----------
    HorizonCode:
      type: string
      description: 1d Tomorrow · 1w 1 week · 1m 1 month · 1y 1 year · 5y 5 years · 10y 10 years · 20y 20 years
      enum: ['1d', '1w', '1m', '1y', '5y', '10y', '20y']
    ScenarioConfiguration:
      type: object
      description: Snapshot of the Scenario Panel. Darkness and Optimism are independent.
      required: [realism, darkness, optimism, horizon, wildcards, customWildcards, output]
      properties:
        realism:
          type: integer
          format: int32
          minimum: 1
          maximum: 10
        darkness:
          type: integer
          format: int32
          minimum: 1
          maximum: 10
        optimism:
          type: integer
          format: int32
          minimum: 1
          maximum: 10
        horizon:
          $ref: '#/components/schemas/HorizonCode'
        wildcards:
          type: array
          description: Enabled catalogue wildcards only (a disabled wildcard is absent). Ids unique.
          maxItems: 30
          items:
            $ref: '#/components/schemas/WildcardSetting'
        customWildcards:
          type: array
          maxItems: 3
          items:
            $ref: '#/components/schemas/CustomWildcard'
        output:
          $ref: '#/components/schemas/OutputSettings'
    WildcardSetting:
      type: object
      required: [wildcardId, intensity]
      properties:
        wildcardId:
          type: string
          minLength: 1
          maxLength: 60
          pattern: '^[a-z0-9-]+$'
          example: biology-new-pandemic
        intensity:
          type: integer
          format: int32
          minimum: 1
          maximum: 10
    CustomWildcard:
      type: object
      required: [label, intensity]
      properties:
        label:
          type: string
          minLength: 1
          maxLength: 40
          pattern: '^\s*\S.*$'
          description: Free text, trimmed by the backend; 1–40 characters after trimming
        intensity:
          type: integer
          format: int32
          minimum: 1
          maximum: 10
    OutputSettings:
      type: object
      required: [story, illustration]
      properties:
        story:
          type: boolean
          description: Must be true in this version
        illustration:
          type: boolean
          description: Must be false in this version (MVP+1)
...
    RunKind:
      type: string
      enum: [STANDARD, ALTERNATIVE]
    RunStatus:
      type: string
      description: QUEUED and RUNNING are active; the others are terminal
      enum: [QUEUED, RUNNING, COMPLETED, INSUFFICIENT_EVIDENCE, FAILED]
    RunStage:
      type: string
      enum:
        - UNDERSTANDING
        - RESEARCH_STRATEGY
        - SEARCHING
        - READING_SOURCES
        - CONNECTING_SIGNALS
        - RANKING
        - EXPLORING_FUTURES
        - CHALLENGING_ASSUMPTIONS
        - CONSTRUCTING_SCENARIO
        - WRITING_STORY
    RunFailureCode:
      type: string
      enum:
        - NEWS_UNAVAILABLE
        - CHATGPT_RATE_LIMITED
        - CHATGPT_UNAVAILABLE
        - CHATGPT_SESSION_EXPIRED
        - RUN_TIMEOUT
        - INVALID_SCENARIO
        - SCENARIO_REJECTED
        - ALTERNATIVE_NOT_DISTINCT
        - INSUFFICIENT_EVIDENCE
        - RUN_INTERRUPTED
        - INTERNAL_ERROR
    RunFailure:
      type: object
      required: [code, message]
      properties:
        code:
          $ref: '#/components/schemas/RunFailureCode'
        message:
          type: string
          description: User-facing message; never technical details
    ResearchCounts:
      type: object
      required: [searches, articlesRetrieved, articlesConsidered, uniqueEvents, eventsSelected, counterSignals, sourcesUsed]
      properties:
        searches:
          type: integer
          format: int32
          minimum: 0
          description: Search queries executed against the news provider
        articlesRetrieved:
          type: integer
          format: int32
          minimum: 0
          description: Raw articles returned by the provider
        articlesConsidered:
          type: integer
          format: int32
          minimum: 0
          description: Articles left after basic filtering (= stored sources)
        uniqueEvents:
          type: integer
          format: int32
          minimum: 0
        eventsSelected:
          type: integer
          format: int32
          minimum: 0
          description: Evidence Pack size (core + supporting + counter-signals)
        counterSignals:
          type: integer
          format: int32
          minimum: 0
        sourcesUsed:
          type: integer
          format: int32
          minimum: 0
          description: Evidence items cited by factsUsed of the final scenario
    GenerationRun:
      type: object
      required: [id, generationId, kind, status, stageIndex, stageCount, configuration, counts, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        generationId:
          type: string
          pattern: '^ORC-\d{4}-\d{2}-\d{2}-\d{4}(-\d+)?$'
          example: ORC-2026-10-02-1842
        kind:
          $ref: '#/components/schemas/RunKind'
        parentRunId:
          type: string
          format: uuid
          description: Set for ALTERNATIVE runs
        status:
          $ref: '#/components/schemas/RunStatus'
        stage:
          $ref: '#/components/schemas/RunStage'
        stageLabel:
          type: string
          example: Searching current events…
        stageIndex:
          type: integer
          format: int32
          minimum: 0
          maximum: 10
          description: 1-based index of the current stage, 0 while QUEUED, 10 when terminal after WRITING_STORY
        stageCount:
          type: integer
          format: int32
          example: 10
        configuration:
          $ref: '#/components/schemas/ScenarioConfiguration'
        evidencePackId:
          type: string
          format: uuid
        counts:
          $ref: '#/components/schemas/ResearchCounts'
        failure:
          $ref: '#/components/schemas/RunFailure'
        suggestedRealism:
          type: integer
          format: int32
          minimum: 1
          maximum: 10
          description: Set when status is INSUFFICIENT_EVIDENCE and realism > 1 (realism − 2, at least 1)
        headline:
          type: string
          description: Set when COMPLETED
        hasOpenCriticIssues:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time

    # ---------- history ----------
    RecentRunList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          maxItems: 20
          items:
            $ref: '#/components/schemas/RecentRunSummary'
    RecentRunSummary:
      type: object
      required: [id, generationId, kind, createdAt, headline, configuration]
      properties:
        id:
          type: string
          format: uuid
        generationId:
          type: string
        kind:
          $ref: '#/components/schemas/RunKind'
        createdAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
        headline:
          type: string
        configuration:
          $ref: '#/components/schemas/ScenarioConfiguration'

    # ---------- research ----------
...